requirements.txt
azure-functions
requests
azure-ai-translation-text
azure-core
azure-data-tables
pytest